refactor(station, line, setting): 패키지 분리
